requirements.txt
Django==3.2.7
djangorestframework==3.12.4
django-environ==0.7.0
django-filter==21.1
django-autoslug==1.9.8
django-countries==7.2.1
pillow==10.2.0
django-phonenumber-field==5.2.0
phonenumbers==8.12.33
psycopg2-binary==2.9.9
psycopg2==2.9.9
flake8==3.9.2
black==18.6b2
isort==5.9.3
djoser==2.1.0
djangorestframework-simplejwt==4.8.0
PyJWT==2.1.0